requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
barcode==1.0.4
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap4==2024.1
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.0.3
django-crispy-forms==2.1
django-filter==23.5
django-pandas==0.6.6
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-widget-tweaks==1.5.0
docopt==0.6.2
et-xmlfile==1.1.0
fontawesomefree==6.5.1
fonttools==4.49.0
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
idna==3.6
Js2Py==0.74
kiwisolver==1.4.5
Levenshtein==0.25.0
lxml==5.1.0
matplotlib==3.8.3
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
openpyxl-image-loader==1.0.5
oscrypto==1.3.0
packaging==23.2
pandas==2.2.1
patsy==0.5.6
phonenumbers==8.13.31
pillow==10.2.0
pipwin==0.5.2
plotly==5.19.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pyHanko==0.22.0
pyhanko-certvalidator==0.26.3
pyjsparser==2.7.1
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.1.0
pypng==0.20220715.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-barcode==0.15.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.6.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.12.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
statsmodels==0.14.1
svglib==1.5.1
tenacity==8.2.3
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
uvicorn==0.30.1
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
